requirements.txt
adns-python==1.2.1
argparse==1.2.1
cffi==0.8.6
dnspython==1.12.0
gevent==1.0.1
greenlet==0.4.5
msgpack-python==0.4.2
publicsuffix==1.0.5
pycares==0.6.3
pythonwhois==2.4.3
pyuv==1.0.1
pyzmq==13.1.0
wsgiref==0.1.2
zerorpc==0.4.4
zope.interface==4.1.1
